+++ src/main/java/com/rackspace/salus/policy/manage/services/PolicyManagement.java
@@ -20,8 +20,8 @@ import com.rackspace.salus.monitor_management.web.client.MonitorApi;
 import com.rackspace.salus.telemetry.model.PolicyScope;
 import com.rackspace.salus.telemetry.entities.MonitorPolicy;
 import com.rackspace.salus.telemetry.entities.Policy;
-import com.rackspace.salus.policy.manage.repositories.MonitorPolicyRepository;
-import com.rackspace.salus.policy.manage.repositories.PolicyRepository;
+import com.rackspace.salus.telemetry.repositories.MonitorPolicyRepository;
+import com.rackspace.salus.telemetry.repositories.PolicyRepository;
 import com.rackspace.salus.policy.manage.web.model.MonitorPolicyCreate;
 import com.rackspace.salus.resource_management.web.client.ResourceApi;
 import com.rackspace.salus.telemetry.errors.AlreadyExistsException;
