+++ src/main/java/com/rackspace/salus/policy/manage/web/model/validator/PolicyValidator.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.rackspace.salus.policy.manage.model.validator;
+package com.rackspace.salus.policy.manage.web.model.validator;
 
 import com.rackspace.salus.telemetry.model.PolicyScope;
 import javax.validation.ConstraintValidator;
